refactor: moved usage command into stats command
