+++ src/commands/stats.ts
@@ -1,18 +1,60 @@
-import { CommandInteraction, Message } from "discord.js";
+import {
+    ActionRowBuilder,
+    ButtonBuilder,
+    ButtonStyle,
+    CommandInteraction,
+    Interaction,
+    InteractionReplyOptions,
+    Message,
+    MessageActionRowComponentBuilder,
+    MessageEditOptions,
+    MessageOptions,
+} from "discord.js";
 import { addCooldown, getResponse, onCooldown } from "../utils/cooldownhandler";
 import prisma from "../utils/database/database";
 import redis from "../utils/database/redis";
 import { getStats } from "../utils/economy/utils";
+import { getCommandUses } from "../utils/karma/utils";
 import { Categories, Command, NypsiCommandInteraction } from "../utils/models/Command";
 import { CustomEmbed } from "../utils/models/EmbedBuilders";
 
-const cmd = new Command("stats", "view your economy stats", Categories.MONEY);
+const cmd = new Command("stats", "view your nypsi stats", Categories.MONEY);
+
+cmd.slashEnabled = true;
+cmd.slashData
+    .addSubcommand((economy) => economy.setName("economy").setDescription("view your economy stats"))
+    .addSubcommand((commands) => commands.setName("commands").setDescription("view your command usage stats"));
 
 async function run(message: Message | (NypsiCommandInteraction & CommandInteraction), args: string[]) {
+    const send = async (data: MessageOptions | InteractionReplyOptions) => {
+        if (!(message instanceof Message)) {
+            if (message.deferred) {
+                await message.editReply(data);
+            } else {
+                await message.reply(data as InteractionReplyOptions);
+            }
+            const replyMsg = await message.fetchReply();
+            if (replyMsg instanceof Message) {
+                return replyMsg;
+            }
+        } else {
+            return await message.channel.send(data as MessageOptions);
+        }
+    };
+
+    const edit = async (data: MessageEditOptions, msg: Message) => {
+        if (!(message instanceof Message)) {
+            await message.editReply(data);
+            return await message.fetchReply();
+        } else {
+            return await msg.edit(data);
+        }
+    };
+
     if (await onCooldown(cmd.name, message.member)) {
         const embed = await getResponse(cmd.name, message.member);
 
-        return message.channel.send({ embeds: [embed] });
+        return send({ embeds: [embed] });
     }
 
     await addCooldown(cmd.name, message.member, 15);
@@ -65,7 +107,146 @@ async function run(message: Message | (NypsiCommandInteraction & CommandInteract
 
         embed.setFooter({ text: `you have performed ${commandUses ? commandUses.toLocaleString() : "0"} commands today` });
 
-        return message.channel.send({ embeds: [embed] });
+        return send({ embeds: [embed] });
+    };
+
+    const commandStats = async () => {
+        const uses = await getCommandUses(message.member);
+
+        const pages = new Map<number, string[]>();
+
+        for (const i of uses) {
+            const str = `\`$${i.command}\` ${i.uses.toLocaleString()}`;
+            if (pages.size == 0) {
+                pages.set(1, [str]);
+            } else if (pages.get(pages.size).length >= 10) {
+                pages.set(pages.size + 1, [str]);
+            } else {
+                const arr = pages.get(pages.size);
+                arr.push(str);
+            }
+        }
+
+        const embed = new CustomEmbed(message.member, pages.get(1).join("\n"))
+            .setHeader("most used commands", message.author.avatarURL())
+            .setFooter({ text: `page 1/${pages.size}` });
+
+        let row = new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(
+            new ButtonBuilder().setCustomId("⬅").setLabel("back").setStyle(ButtonStyle.Primary).setDisabled(true),
+            new ButtonBuilder().setCustomId("➡").setLabel("next").setStyle(ButtonStyle.Primary)
+        );
+
+        let msg: Message;
+
+        if (pages.size == 1) {
+            return await send({ embeds: [embed] });
+        } else {
+            msg = await send({ embeds: [embed], components: [row] });
+        }
+
+        const filter = (i: Interaction) => i.user.id == message.author.id;
+
+        let currentPage = 1;
+
+        const pageManager = async (): Promise<void> => {
+            const reaction = await msg
+                .awaitMessageComponent({ filter, time: 30000 })
+                .then(async (collected) => {
+                    await collected.deferUpdate();
+                    return collected.customId;
+                })
+                .catch(async () => {
+                    await edit({ components: [] }, msg).catch(() => {});
+                });
+
+            if (!reaction) return;
+
+            const newEmbed = new CustomEmbed(message.member).setHeader("most used commands", message.author.avatarURL());
+
+            if (reaction == "⬅") {
+                if (currentPage <= 1) {
+                    return pageManager();
+                } else {
+                    currentPage--;
+
+                    newEmbed.setDescription(pages.get(currentPage).join("\n"));
+
+                    newEmbed.setFooter({ text: `page ${currentPage}/${pages.size}` });
+
+                    if (currentPage == 1) {
+                        row = new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(
+                            new ButtonBuilder()
+                                .setCustomId("⬅")
+                                .setLabel("back")
+                                .setStyle(ButtonStyle.Primary)
+                                .setDisabled(true),
+                            new ButtonBuilder()
+                                .setCustomId("➡")
+                                .setLabel("next")
+                                .setStyle(ButtonStyle.Primary)
+                                .setDisabled(false)
+                        );
+                    } else {
+                        row = new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(
+                            new ButtonBuilder()
+                                .setCustomId("⬅")
+                                .setLabel("back")
+                                .setStyle(ButtonStyle.Primary)
+                                .setDisabled(false),
+                            new ButtonBuilder()
+                                .setCustomId("➡")
+                                .setLabel("next")
+                                .setStyle(ButtonStyle.Primary)
+                                .setDisabled(false)
+                        );
+                    }
+                    await edit({ embeds: [newEmbed], components: [row] }, msg);
+                    return pageManager();
+                }
+            } else if (reaction == "➡") {
+                if (currentPage >= pages.size) {
+                    return pageManager();
+                } else {
+                    currentPage++;
+
+                    newEmbed.setDescription(pages.get(currentPage).join("\n"));
+
+                    newEmbed.setFooter({ text: `page ${currentPage}/${pages.size}` });
+
+                    if (currentPage == pages.size) {
+                        row = new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(
+                            new ButtonBuilder()
+                                .setCustomId("⬅")
+                                .setLabel("back")
+                                .setStyle(ButtonStyle.Primary)
+                                .setDisabled(false),
+                            new ButtonBuilder()
+                                .setCustomId("➡")
+                                .setLabel("next")
+                                .setStyle(ButtonStyle.Primary)
+                                .setDisabled(true)
+                        );
+                    } else {
+                        row = new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(
+                            new ButtonBuilder()
+                                .setCustomId("⬅")
+                                .setLabel("back")
+                                .setStyle(ButtonStyle.Primary)
+                                .setDisabled(false),
+                            new ButtonBuilder()
+                                .setCustomId("➡")
+                                .setLabel("next")
+                                .setStyle(ButtonStyle.Primary)
+                                .setDisabled(false)
+                        );
+                    }
+                    await edit({ embeds: [newEmbed], components: [row] }, msg);
+                    return pageManager();
+                }
+            }
+        };
+
+        return pageManager();
     };
 
     if (args.length == 0) {
@@ -200,7 +381,15 @@ async function run(message: Message | (NypsiCommandInteraction & CommandInteract
         });
 
         embed.setHeader("global stats", message.author.avatarURL());
-        return message.channel.send({ embeds: [embed] });
+        return send({ embeds: [embed] });
+    } else if (
+        args[0].toLowerCase() == "economy" ||
+        args[0].toLowerCase() == "gamble" ||
+        args[0].toLowerCase().includes("item")
+    ) {
+        return normalStats();
+    } else if (args[0].toLowerCase().includes("command") || args[0].toLowerCase().includes("cmd")) {
+        return commandStats();
     } else {
         return normalStats();
     }
